oil_app/requirements.txt
streamlit>=1.8.1
streamlit-folium==0.6.15
pandas==1.3.4
Pillow==7.0.0
haversine==2.7.0
geopandas==0.11.1
geojson==2.5.0
requests==2.22.0
requests-html==0.10.0